Python reimplementation of the autosegmenter code
This avoids the dependency on the specific version of the Intel
Performance Primitives library.  The code tries to follow the original
C++ autosegmentation but deviates in a few places. Based on a couple
of random tests the segmentations are mostly similar.

Diamond filter functionality is not working quite yet.

As there is overlap in places with the gemini/delphi parts of
dermshare, such as the used python libraries and provided
functionality, this should possibly get folded into the existing
gemini/delphi codebase.